BE: [Fix} Refactoring (#80)
* [ADD] be/seungil branch

* [FIX] fix to restart

* [ADD] new structure

* [UPD] structure

* [Feat] create, read of docs

* [UPD] Patch, Delete of Docs Table

* [Fix] TODO 추가

* [Feat] added video streaming

* [TODO] add comments of todo

* [ADD] carousel

* [ADD] config for s3 bucket

* [Feat] S3 Uploader

* [Feat] Carousel read

* [UPD] s3 upload carousel

* [Feat] Create & Delete of Chapter

* [Feat] upload, delete of Chapter

* [FIX] mysql 예약어

* conflict 해결

* [Fix] field name

* [Fix] Carousel upload

* [Feat] create, delete of contents

* [UPD] contents-chapter-uploadClass-docs Connection  & video upload: byte -> s3

* [Feat] Delete UploadClass & disclosure of Contents

* [Feat] UploadClass CRUD 완성

* [Feat] CurriculumInfo

* [Del] workspace.xml

* [Feat] 콘텐츠 상세화면

* [ADD] Review feat

* [Fix] CascadeType

* [Fix] Docs Exception

* [Fix] UploadClass Exception

* [Fix] Chapter Exception

* [Feat] Video Streaming ResponseDto

* [Feat] Warning save Handler

* [Fix] ChapterService Exception

* [Feat] Wanring Create & Read in myPage

* [Feat] Bookmark CRUD

* [ADD] PageInfo & MultiResponseDto

* [Feat] Search for Content Title Order By Lateast

* [Feat] Search Engine

* [Fix] 빌드시 충돌 체크

* [Fix] TODO

* [Feat] Google Social

* [UPD] service logic

* [Feat] User Social Login

* [Fix] refactoring

* [Fix] application.yml & build.gradle

* [Feat] Login

* [ADD] CORS Configuration

* [Fix] CORS: SecurityConfig

* [ADD] Enum Converter

* [UPD] working tree update

* [Fix] endPoints

* [UPD] contents Avg:starRate

* [UPD] disclosure 노출 여부

* [Feat] 콘텐츠 구매 여부 포함 DTO

* [FIx] CRUD 동작 권한 수정

* [DEL] uploadClass - constant

* [Fix] docs exists logic del

* [Fix] Docs Transaction 수정

* [UPD] 콘텐츠 상세화면 wished 추가

* [UPD] refactoring

* [Fix] bookmark, review DTO field 추가

* [Fix] Refactoring 주 기능

+++ server/BED/src/main/java/main008/BED/contents/service/ContentsService.java
@@ -3,6 +3,7 @@ package main008.BED.contents.service;
 import lombok.RequiredArgsConstructor;
 import main008.BED.S3.S3ServiceImpl;
 import main008.BED.chapter.entity.Chapter;
+import main008.BED.contents.dto.ContentsDto;
 import main008.BED.contents.entity.Contents;
 import main008.BED.contents.repository.ContentsRepository;
 import main008.BED.exception.BusinessLogicException;
@@ -16,6 +17,8 @@ import main008.BED.payment.entity.Payment;
 import main008.BED.payment.service.PaymentService;
 import main008.BED.review.entity.Review;
 import main008.BED.uploadClass.entity.UploadClass;
+import main008.BED.users.entity.Users;
+import main008.BED.users.repository.UsersRepository;
 import main008.BED.users.service.UsersService;
 import main008.BED.wish.entity.Wish;
 import main008.BED.wish.service.WishService;
@@ -24,6 +27,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -35,6 +39,7 @@ public class ContentsService {
 
     private final MyUploadClassService myUploadClassService;
     private final ContentsRepository contentsRepository;
+    private final UsersRepository usersRepository;
     private final PaymentService paymentService;
     private final MyClassService myClassService;
     private final S3ServiceImpl s3ServiceImpl;
@@ -73,11 +78,17 @@ public class ContentsService {
     /**
      * contents 삭제
      */
-    public void removeContents(Long contentId) {
+    public void removeContents(Long contentId, Principal principal) {
 
         // TODO: 컨텐츠에 담긴 챕터, DOCS 연쇄 삭제
         Contents byContentsId = contentsRepository.findByContentsId(contentId).orElseThrow(()
                 -> new BusinessLogicException(ExceptionCode.CONTENTS_NOT_FOUND));
+
+        Users byEmail = usersRepository.findByEmail(principal.getName());
+
+        if (byContentsId.getUsers().getUsersId() != byEmail.getUsersId()) {
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_DELETE);
+        }
 
         String fileKey = byContentsId.getFileKey();
         s3ServiceImpl.delete(fileKey, "/contents/thumbnail");
@@ -218,18 +229,7 @@ public class ContentsService {
 
         return contentsPage;
     }
-    
-    private static ArrayList<Contents> getDiscloseContents(List<Contents> contentsList) {
-        ArrayList<Contents> discloseList = new ArrayList<>();
 
-        for (Contents content : contentsList) {
-            Boolean disclosure = content.getDisclosure();
-            if (disclosure) {
-                discloseList.add(content);
-            }
-        }
-        return discloseList;
-    }
 
     /**
      * 총 별점 계산
@@ -267,6 +267,41 @@ public class ContentsService {
             float avg = (float) sum / count;
             return Math.round(avg * 100) / 100.0; // 소숫점 둘 째 자리까지
         }
+
+    }
+
+    private static ArrayList<Contents> getDiscloseContents(List<Contents> contentsList) {
+        ArrayList<Contents> discloseList = new ArrayList<>();
+
+        for (Contents content : contentsList) {
+            Boolean disclosure = content.getDisclosure();
+            if (disclosure) {
+                discloseList.add(content);
+            }
+        }
+        return discloseList;
+    }
+
+    /**
+     * UPDATE: Contents 업데이트
+     */
+    public void updateContents(Long contentsId, Principal principal, Contents newContents, Payment newPayment) {
+        Contents oldContents = contentsRepository.findByContentsId(contentsId).orElseThrow(()
+                -> new BusinessLogicException(ExceptionCode.CONTENTS_NOT_FOUND));
+
+        Users reqUser = usersRepository.findByEmail(principal.getName());
+
+        if (oldContents.getUsers().getUsersId() != reqUser.getUsersId()) {
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_PATCH);
+        }
+
+        oldContents.setCategories(newContents.getCategories());
+        oldContents.setDetails(newContents.getDetails());
+        oldContents.setTitle(newContents.getTitle());
+        oldContents.setThumbnail(newContents.getThumbnail());
+        oldContents.setTutorDetail(newContents.getTutorDetail());
+        Payment oldPayment = oldContents.getPayment();
+        oldPayment.setPrice(newPayment.getPrice());
 
     }
